+++ packages/riverpod_generator/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased patch
+
+- Fix all `Notifier.overrideWith` throwing a cast error.
+
 ## 2.2.6 - 2023-08-16
 
 - Support both analyzer 5.12.0 and analyzer 6.0.0