+++ packages/riverpod_generator/CHANGELOG.md
@@ -1,5 +1,17 @@
-## Unreleased patch
+## Unreleased minor
 
+- The "ref" object now contains the provider parameters too.
+  This enabled `provider.overrideWith` to use the provider arguments:
+  ```dart
+  @riverpod
+  int example(ExampleRef ref, {int? id}) { /* */ }
+  // ...
+  exampleProvider.overrideWith(
+    (ref) {
+      print(ref.id);
+    }
+  )
+  ```
 - Fix all `provider.overrideWith` causing a cast error if the notifier
   receives arguments.